+++ GridGraph/write.h
@@ -4,6 +4,6 @@
 #include <stdio.h>
 
 // FUNCTIONS DECLARATIONS
-int write(FILE* file, int width, int height, float edge_weight_min, float edge_weight_max, int edge_count_min, int edge_count_max);
+int write(FILE* file, int width, int height, double edge_weight_min, double edge_weight_max, int edge_count_min, int edge_count_max);
 
 #endif
